refactor(storage): drop assistant-auth and assistant-backup dependencies
Completes the storage→auth/backup dependency inversion:

- Move `backup_legacy` to `assistant_backup::backup_legacy_install`.
- Move `create_admin_user` + `AdminCredentials` to
  `assistant_auth::bootstrap`. Takes `&dyn UserStore` + `&dyn
  MembershipStore` so the auth crate doesn't gain a storage dep.
- Refactor `storage::migration::migrate_database` to return the opened
  `OrgStorageLayer` plus the seeded `OrgId`/`SpaceId` so the caller can
  layer admin bootstrap on top.
- Remove `run_migration` orchestrator. Production callers (web-ui main)
  now compose the four steps explicitly: backup → filesystem migration
  → database migration → admin bootstrap.
- Drop `assistant-auth`, `assistant-backup`, and the now-unused `rand`
  from `crates/storage/Cargo.toml [dependencies]`. Auth/backup remain as
  `[dev-dependencies]` so the migration round-trip test mirrors the
  production pipeline.
- Add `assistant-backup` as a direct dep of `assistant-web-ui` (it now
  calls the backup helper directly).
- Document the dep-boundary contract in `crates/storage/src/lib.rs`.

Storage now depends only on `assistant-core` (and `assistant-skills` for
the registry helpers) in production. The `dep_boundary` integration test
enforces this at compile time.
